+++ packages/core/e2e/order.e2e-spec.ts
@@ -17,7 +17,7 @@ import gql from 'graphql-tag';
 import path from 'path';
 
 import { initialData } from '../../../e2e-common/e2e-initial-data';
-import { TEST_SETUP_TIMEOUT_MS, testConfig } from '../../../e2e-common/test-config';
+import { testConfig, TEST_SETUP_TIMEOUT_MS } from '../../../e2e-common/test-config';
 
 import {
     failsToSettlePaymentMethod,
@@ -83,9 +83,9 @@ import {
     DELETE_SHIPPING_METHOD,
     GET_CUSTOMER_LIST,
     GET_ORDER,
+    GET_ORDERS_LIST,
     GET_ORDER_FULFILLMENTS,
     GET_ORDER_HISTORY,
-    GET_ORDERS_LIST,
     GET_PRODUCT_WITH_VARIANTS,
     GET_STOCK_MOVEMENT,
     SETTLE_PAYMENT,
@@ -275,6 +275,30 @@ describe('Orders resolver', () => {
             ]);
         });
 
+        it('sort by customerLastName', async () => {
+            async function sortOrdersByLastName(sortOrder: SortOrder) {
+                const { orders } = await adminClient.query<GetOrderList.Query, GetOrderList.Variables>(
+                    GET_ORDERS_LIST,
+                    {
+                        options: {
+                            sort: {
+                                customerLastName: sortOrder,
+                            },
+                        },
+                    },
+                );
+                return orders;
+            }
+
+            const result1 = await sortOrdersByLastName(SortOrder.ASC);
+            expect(result1.totalItems).toEqual(2);
+            expect(result1.items.map(order => order.customer?.lastName)).toEqual(['Donnelly', 'Zieme']);
+
+            const result2 = await sortOrdersByLastName(SortOrder.DESC);
+            expect(result2.totalItems).toEqual(2);
+            expect(result2.items.map(order => order.customer?.lastName)).toEqual(['Zieme', 'Donnelly']);
+        });
+
         it('filter by total', async () => {
             const result = await adminClient.query<GetOrderList.Query, GetOrderList.Variables>(
                 GET_ORDERS_LIST,
@@ -324,6 +348,23 @@ describe('Orders resolver', () => {
                 { id: 'T_2', totalQuantity: 4 },
             ]);
         });
+
+        it('filter by customerLastName', async () => {
+            const result = await adminClient.query<GetOrderList.Query, GetOrderList.Variables>(
+                GET_ORDERS_LIST,
+                {
+                    options: {
+                        filter: {
+                            customerLastName: {
+                                eq: customers[1].lastName,
+                            },
+                        },
+                    },
+                },
+            );
+            expect(result.orders.totalItems).toEqual(1);
+            expect(result.orders.items[0].customer?.lastName).toEqual(customers[1].lastName);
+        });
     });
 
     describe('payments', () => {
@@ -356,6 +397,7 @@ describe('Orders resolver', () => {
             });
 
             expect(result.order!.state).toBe('PaymentAuthorized');
+            expect(result.order!.payments![0].state).toBe('Cancelled');
             firstOrderCode = order.code;
             firstOrderId = order.id;
         });