+++ packages/elasticsearch-plugin/src/build-elastic-body.ts
@@ -1,6 +1,8 @@
 import { LanguageCode, LogicalOperator, PriceRange, SortOrder } from '@vendure/common/lib/generated-types';
 import { DeepRequired, ID } from '@vendure/core';
 
+import { UserInputError } from '../../core/src/common/error/errors';
+
 import { SearchConfig } from './options';
 import { ElasticSearchInput, SearchRequestBody } from './types';
 
@@ -26,6 +28,7 @@ export function buildElasticBody(
         sort,
         priceRangeWithTax,
         priceRange,
+        facetValueFilters,
     } = input;
     const query: any = {
         bool: {},
@@ -58,6 +61,24 @@ export function buildElasticBody(
                 bool: { [operator]: facetValueIds.map(id => ({ term: { facetValueIds: id } })) },
             },
         ]);
+    }
+    if (facetValueFilters && facetValueFilters.length) {
+        ensureBoolFilterExists(query);
+        facetValueFilters.forEach(facetValueFilter => {
+            if (facetValueFilter.and && facetValueFilter.or && facetValueFilter.or.length) {
+                throw new UserInputError('error.facetfilterinput-invalid-input');
+            }
+
+            if (facetValueFilter.and) {
+                query.bool.filter.push({ term: { facetValueIds: facetValueFilter.and } });
+            }
+
+            if (facetValueFilter.or && facetValueFilter.or.length) {
+                query.bool.filter.push({
+                    bool: { ['should']: facetValueFilter.or.map(id => ({ term: { facetValueIds: id } })) },
+                });
+            }
+        });
     }
     if (collectionId) {
         ensureBoolFilterExists(query);